Add support for interactive learning
