+++ rasa/core/channels/console.py
@@ -10,9 +10,11 @@ from async_generator import async_generator, yield_
 from prompt_toolkit.styles import Style
 
 import rasa.cli.utils as cliutils
+from rasa.cli.utils import button_to_string, element_to_string
+
 from rasa.core import utils
 from rasa.core.channels.channel import UserMessage
-from rasa.core.channels.channel import RestInput, button_to_string, element_to_string
+from rasa.core.channels.channel import RestInput
 from rasa.core.constants import DEFAULT_SERVER_URL
 from rasa.core.interpreter import INTENT_MESSAGE_PREFIX
